refactor(banners): moving banner components outside of the merchandising folder
