+++ packages/snap-preact/components/src/components/Atoms/Banner/Banner.tsx
@@ -3,12 +3,12 @@ import { observer } from 'mobx-react-lite';
 import { jsx, css } from '@emotion/react';
 import classnames from 'classnames';
 
-import { Theme, useTheme, CacheProvider, useTreePath, withTracking, withController } from '../../../../providers';
-import { mergeProps, mergeStyles } from '../../../../utilities';
+import { Theme, useTheme, CacheProvider, useTreePath, withTracking, withController } from '../../../providers';
+import { mergeProps, mergeStyles } from '../../../utilities';
 
 import { BannerContent, ContentType } from '@searchspring/snap-store-mobx';
 import type { AutocompleteController, SearchController } from '@searchspring/snap-controller';
-import { ComponentProps, StyleScript } from '../../../../types';
+import { ComponentProps, StyleScript } from '../../../types';
 import { useCallback } from 'preact/hooks';
 
 const defaultStyles: StyleScript<BannerProps> = () => {
