+++ packages/snap-preact/components/src/components/Atoms/Banner/index.ts
