+++ packages/snap-preact/components/src/components/Atoms/InlineBanner/index.ts
