+++ packages/snap-preact/components/src/components/Atoms/InlineBanner/readme.md
