--- packages/snap-preact/components/src/components/Atoms/Merchandising/index.ts
@@ -1,2 +1,0 @@
-export * from './Banner';
-export * from './InlineBanner';
