+++ packages/snap-preact/components/src/components/Organisms/Autocomplete/Autocomplete.tsx
@@ -12,7 +12,7 @@ import type { Term } from '@searchspring/snap-store-mobx';
 
 import { Icon, IconProps } from '../../Atoms/Icon/Icon';
 import { Results, ResultsProps } from '../Results';
-import { Banner, BannerProps } from '../../Atoms/Merchandising/Banner';
+import { Banner, BannerProps } from '../../Atoms/Banner';
 import { Facets, FacetsProps } from '../Facets';
 import { defined, cloneWithProps, mergeProps, mergeStyles } from '../../../utilities';
 import { createHoverProps } from '../../../toolbox';
