+++ packages/snap-preact/components/src/index.ts
@@ -3,14 +3,15 @@ export * from './components/Atoms/BadgeImage';
 export * from './components/Atoms/BadgePill';
 export * from './components/Atoms/BadgeRectangle';
 export * from './components/Atoms/BadgeText';
+export * from './components/Atoms/Banner';
 export * from './components/Atoms/Breadcrumbs';
 export * from './components/Atoms/Button';
 export * from './components/Atoms/Dropdown';
 export * from './components/Atoms/FormattedNumber';
 export * from './components/Atoms/Icon';
 export * from './components/Atoms/Image';
+export * from './components/Atoms/InlineBanner';
 export * from './components/Atoms/Loading';
-export * from './components/Atoms/Merchandising';
 export * from './components/Atoms/Overlay';
 export * from './components/Atoms/PaginationInfo';
 export * from './components/Atoms/Price';
