+++ packages/snap-preact/components/src/providers/themeComponents.ts
@@ -10,7 +10,8 @@ import type { FormattedNumberProps } from '../components/Atoms/FormattedNumber';
 import type { IconProps, IconNames } from '../components/Atoms/Icon';
 import type { ImageProps } from '../components/Atoms/Image';
 import type { LoadingBarProps } from '../components/Atoms/Loading';
-import type { BannerNames, BannerProps, InlineBannerProps } from '../components/Atoms/Merchandising';
+import type { BannerNames, BannerProps } from '../components/Atoms/Banner';
+import type { InlineBannerProps } from '../components/Atoms/InlineBanner';
 import type { OverlayProps } from '../components/Atoms/Overlay';
 import type { PaginationInfoProps } from '../components/Atoms/PaginationInfo';
 import type { PriceNames, PriceProps } from '../components/Atoms/Price';
